pyproject.toml
[project]
name = "ttb-label-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "monsterui>=1.0.29",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-fasthtml>=0.12.29",
    "ruff>=0.14.1",
]
